experimental fix for problem reported by Zhe Zhang; file system with a large device id value shows up as UNKNOWN in the parser output, likely because of a type casting mismatch in the header and file record
git-svn-id: https://svn.mcs.anl.gov/repos/darshan/branches/darshan-1.1@254 3b7491f3-a168-0410-bf4b-c445ed680a29

+++ darshan-logutils.h
@@ -18,7 +18,7 @@ int darshan_log_getfile(darshan_fd fd,
     struct darshan_job* job, 
     struct darshan_file *file);
 int darshan_log_getexe(darshan_fd fd, char *buf, int *flag);
-int darshan_log_getmounts(darshan_fd fd, int** devs, char*** mnt_pts, char***
+int darshan_log_getmounts(darshan_fd fd, int64_t** devs, char*** mnt_pts, char***
     fs_types, int* count, int *flag);
 void darshan_log_close(darshan_fd file);
 void darshan_log_print_version_warnings(struct darshan_job *job);
